WIP refactor

+++ include/geometry/shpere/sphere.h
@@ -4,6 +4,6 @@
 typedef struct Vulkan Vulkan;
 typedef unsigned int uint32_t;
 
-void makeSphere(Vulkan *, uint32_t, uint32_t, uint32_t);
+void makeSphere(Vulkan *, uint32_t, uint32_t, float);
 
 #endif /* INCLUDE_GEOMETRY_SHPERE_SPHERE */
